myproject/requirements.txt
Django==1.11.10
mysqlclient==1.3.12
pytz==2018.3
click==6.7
client==0.0.1
Django==1.11.10
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
MarkupSafe==1.0
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.12
Pygments==2.2.0
pytz==2018.3
virtualenv==15.1.0


